% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{gtable_arrange}
\alias{gtable_arrange}
\title{Arrange multiple grobs on a page}
\usage{
gtable_arrange(..., grobs = list(), as.table = TRUE, top = NULL,
  bottom = NULL, left = NULL, right = NULL, draw = TRUE)
}
\arguments{
\item{...}{grobs to arrange, and optional layout parameters}

\item{grobs}{[optional] list of grobs}

\item{as.table}{logical, indicates whether the list of grobs should be placed row-wise or column-wise}

\item{top}{annotation (text or grob)}

\item{bottom}{annotation (text or grob)}

\item{left}{annotation (text or grob)}

\item{right}{annotation (text or grob)}

\item{draw}{logical, draw the table}
}
\description{
Arrange multiple grobs on a page
}

