% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.R
\name{as.gtable}
\alias{as.gtable}
\alias{as.gtable.grob}
\title{Convert to a gtable}
\usage{
as.gtable(x, ...)

\method{as.gtable}{grob}(x, widths = NULL, heights = NULL, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Arguments forwarded to methods.}

\item{widths, heights}{Scalar unit setting the size of the table. Defaults
to \code{\link[grid:grobWidth]{grid::grobWidth()}} and \code{\link[grid:grobWidth]{grid::grobHeight()}} of \code{x} respectively.}
}
\value{
A gtable object
}
\description{
Convert to a gtable
}
\section{Methods (by class)}{
\itemize{
\item \code{as.gtable(grob)}: Creates a 1-cell gtable containing the grob.

}}
