% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable-package.R
\docType{package}
\name{gtable-package}
\title{gtable: Arrange 'Grobs' in Tables}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Tools to make it easier to work with "tables" of
    'grobs'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/gtable}
  \item Report bugs at \url{https://github.com/r-lib/gtable/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
