\name{gtable_add_cols}
\alias{gtable_add_cols}
\title{Add new columns in specified position.}
\usage{
  gtable_add_cols(x, widths, clip = "inherit", pos = -1)
}
\arguments{
  \item{pos}{new row will be added below this position.
  Defaults to adding col on right. \code{0} adds on the
  left.}
}
\description{
  Add new columns in specified position.
}

