\name{gtable_add_rows}
\alias{gtable_add_rows}
\title{Add new rows in specified position.}
\usage{
  gtable_add_rows(x, heights, clip = "inherit", pos = -1)
}
\arguments{
  \item{pos}{new row will be added below this position.
  Defaults to adding row on bottom. \code{0} adds on the
  top.}
}
\description{
  Add new rows in specified position.
}

