% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.r
\name{gtable_height}
\alias{gtable_height}
\title{Returns the height of a gtable, in the gtable's units}
\usage{
gtable_height(x)
}
\arguments{
\item{x}{A gtable object}
}
\description{
Note that unlike heightDetails.gtable, this can return relative units.
}
