% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-grob.R
\name{gtable_remove_grob}
\alias{gtable_remove_grob}
\title{Remove a grob from a gtable}
\usage{
gtable_remove_grob(x, grob)
}
\arguments{
\item{x}{A gtable object}

\item{grob}{Either a character or integer vector identifying grobs to remove}
}
\value{
A gtable without the specified grobs
}
\description{
This function takes grob indexes or names and remove them from the gtable.
}
