% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gtable_table}
\alias{gtable_table}
\title{Graphical display of a textual table.}
\usage{
gtable_table(d, colour = "black", fill = NA)
}
\arguments{
  \item{d}{data.frame or matrix}

  \item{colour}{label colours}

  \item{fill}{cell background colours}
}
\description{
Graphical display of a textual table.
}
\examples{
\donttest{
d <- head(iris, 3)
core <- gtable_table(d, 1:5)
colhead <- gtable_table(t(colnames(d)))
rowhead <- gtable_table(c("", rownames(d)))
g <- rbind(colhead, core)
g <- cbind(rowhead, g)
grid.newpage()
grid.draw(g)
}
}

