% Generated by roxygen2 (4.0.0): do not edit by hand
\name{gtable_from_layout}
\alias{gtable_from_layout}
\title{Arrange multiple grobs on a page following a layout matrix}
\usage{
gtable_from_layout(grobs, mat, widths = NULL, heights = NULL, ...)
}
\arguments{
  \item{...}{additional parameters passed to gtable}

  \item{grobs}{list of grobs}

  \item{widths}{}

  \item{heights}{}

  \item{mat}{integer matrix defining the layout}
}
\description{
Arrange multiple grobs on a page following a layout matrix
}

