\name{z_normalize}
\alias{z_normalize}
\title{Normalize z values within a gtable object}
\usage{
  z_normalize(x)
}
\arguments{
  \item{x}{A gtable object}
}
\description{
  The z values within a gtable object can be any numeric
  values. This function will change them to integers
  (starting from 1), preserving the original order.
}

