% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable-package.R
\docType{package}
\name{gtable-package}
\title{gtable: Arrange 'Grobs' in Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to make it easier to work with "tables" of 'grobs'. The 'gtable' package defines a 'gtable' grob class that specifies a grid along with a list of grobs and their placement in the grid. Further the package makes it easy to manipulate and combine 'gtable' objects so that complex compositions can be built up sequentially.
}
\seealso{
\code{\link[=gtable]{gtable()}} for an overview of the class and construction
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
