% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable-package.R
\docType{package}
\name{gtable-package}
\title{gtable: Arrange 'Grobs' in Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to make it easier to work with "tables" of 'grobs'. The 'gtable' package defines a 'gtable' grob class that specifies a grid along with a list of grobs and their placement in the grid. Further the package makes it easy to manipulate and combine 'gtable' objects so that complex compositions can be built up sequentially.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gtable.r-lib.org}
  \item \url{https://github.com/r-lib/gtable}
  \item Report bugs at \url{https://github.com/r-lib/gtable/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder]
}

}
\keyword{internal}
