\name{gtable_add_padding}
\alias{gtable_add_padding}
\title{Add padding around edges of table.}
\usage{
  gtable_add_padding(x, padding)
}
\arguments{
  \item{x}{a \code{\link{gtable}} object}

  \item{padding}{vector of length 4: top, right, bottom,
  left.  Normal recycling rules apply.}
}
\description{
  Add padding around edges of table.
}

