% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable-layouts.R
\name{gtable_col}
\alias{gtable_col}
\title{Create a single column gtable}
\usage{
gtable_col(name, grobs, width = NULL, heights = NULL, z = NULL, vp = NULL)
}
\arguments{
\item{name}{a string giving the name of the table. This is used to name
the layout viewport}

\item{grobs}{a single grob or a list of grobs}

\item{width}{a unit vector giving the width of this column}

\item{heights}{a unit vector giving the height of each row}

\item{z}{a numeric vector giving the order in which the grobs should be
plotted.  Use \code{Inf} (the default) to plot above or \code{-Inf}
below all existing grobs. By default positions are on the integers,
giving plenty of room to insert new grobs between existing grobs.}

\item{vp}{a grid viewport object (or NULL).}
}
\value{
A gtable with one column and as many rows as elements in the grobs
list.
}
\description{
This function stacks a list of grobs into a single column gtable of the given
width and heights.
}
\examples{
library(grid)
a <- rectGrob(gp = gpar(fill = "red"))
b <- circleGrob()
c <- linesGrob()
gt <- gtable_col("demo", list(a, b, c))
gt
plot(gt)
gtable_show_layout(gt)
}
\seealso{
Other gtable construction: 
\code{\link{gtable}()},
\code{\link{gtable_matrix}()},
\code{\link{gtable_row}()},
\code{\link{gtable_spacer}}
}
\concept{gtable construction}
