\name{gtable_gTree}
\alias{gtable_gTree}
\title{Generate a grob (gTree) to render a gtable.}
\usage{
  gtable_gTree(x, ...)
}
\arguments{
  \item{x}{a \code{\link{gtable}} object}

  \item{...}{other parameters passed on to
  \code{\link[grid]{gTree}}}
}
\description{
  Generate a grob (gTree) to render a gtable.
}

