% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grid.r
\name{gtable_show_layout}
\alias{gtable_show_layout}
\title{Visualise the layout of a gtable.}
\usage{
gtable_show_layout(x)
}
\arguments{
\item{x}{a gtable object}
}
\description{
Visualise the layout of a gtable.
}

