\name{gtable_spacer}
\alias{gtable_col_spacer}
\alias{gtable_row_spacer}
\alias{gtable_spacer}
\title{Create a row/col spacer gtable.}
\usage{
  gtable_row_spacer(widths)

  gtable_col_spacer(heights)
}
\arguments{
  \item{widths}{unit vector of widths}

  \item{heights}{unit vector of heights}
}
\description{
  Create a row/col spacer gtable.
}

