% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable-layouts.R
\name{gtable_spacer}
\alias{gtable_spacer}
\alias{gtable_row_spacer}
\alias{gtable_col_spacer}
\title{Create a row/col spacer gtable.}
\usage{
gtable_row_spacer(widths)

gtable_col_spacer(heights)
}
\arguments{
\item{widths}{unit vector of widths}

\item{heights}{unit vector of heights}
}
\value{
A gtable object
}
\description{
Create a zero-column or zero-row gtable with the given heights or widths
respectively.
}
\seealso{
Other gtable construction: 
\code{\link{gtable}()},
\code{\link{gtable_col}()},
\code{\link{gtable_matrix}()},
\code{\link{gtable_row}()}
}
\concept{gtable construction}
