% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtable.R
\name{gtable_width}
\alias{gtable_width}
\title{Returns the width of a gtable, in the gtable's units}
\usage{
gtable_width(x)
}
\arguments{
\item{x}{A gtable object}
}
\description{
Note that unlike widthDetails.gtable, this can return relative units.
}
